+++ apps/docs/components/mdx/components.tsx
@@ -2,6 +2,7 @@ import type { HTMLAttributes, DetailedHTMLProps } from "react";
 import dynamic from "next/dynamic";
 
 import Card from "@/components/card/Card.tsx";
+import Callout from "@/components/callout/Callout.tsx";
 import NextImage from "@/components/image/Image.tsx";
 import Pre from "@/components/pre/Pre.tsx";
 import InlineCode from "@/components/code/InlineCode.tsx";
@@ -33,6 +34,7 @@ type HeadingProps = DetailedHTMLProps<HTMLAttributes<HTMLHeadingElement>, HTMLHe
 export const components = {
     Card,
     code: InlineCode,
+    Callout: Callout,
     Image: NextImage,
     pre: Pre,
     MotionPreview: MotionPreview,